+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Change log for releases.
 
 * Propagate `kaitai_compiler` messages as part of plugin's error message
 * Configure timeout for the execution operations
+* Configure custom KaitaiStream implementations for static builder `fromFile(...)`
 
 # 0.1.2